update jenkins shared lib 	modified:   checkLoadExists.groovy 	new file:   determineDateValue.groovy
