+++ jenkins/vars/determineDateValue.groovy
@@ -1,0 +1,16 @@
+import java.text.SimpleDateFormat
+
+def call() {
+    date = new Date()
+    sdate = new SimpleDateFormat("yyy-MM-dd")
+    Calendar cal = Calendar.getInstance()
+                
+    if(params.BuildDate == "today") {
+        dateValue = sdate.format(date)
+    } else {
+        dateValue = params.BuildDate
+    }
+    echo dateValue
+
+    return dateValue
+}
